backend/requirements.txt
httpx
pandas
numpy
python-dotenv
fastapi
uvicorn[standard]
pydantic
psycopg2-binary
sqlalchemy
alembic
passlib==1.7.4
bcrypt==4.0.1
paho-mqtt
python-jose[cryptography] 
python-multipart
APScheduler